Make sending and receiving async
Before this, we would only send an update to clients after they sent us one.

Now we send updates to clients as soon as any client updates their position.
This can still be improved further, as we don't always need to send an update, and JSON is still inefficient. We probably should be using EVENT_WRITE too.